Implemented weapon functionality, weapons spawn from the players position and travel in a circular arc in the direction of the moouse (approximetley) - currenlt the angle function is somewhat hard coded and may vary between machines and screens.
Weapons have the implemented functionallty to rotate forming an arching
swing but this however is currently commented out as the render engine
does not allow for rotating images. - Likly to be replaced with an
animation system.

How weapons function:
Weapon component acts as a component that is put onto the player allow
it to 'weild' a weapon.

Upon swinging (pressing left click) an event "playerAttack" will trigger
on the player and this will active the WeaponComponent class

The weaponpon component will generate a new "project" entity (meele
weapons currently function as projectile), this new weapon will be
spawned at the players position using the PlaceStructreAt service
(credit to the Buidlables teams). This was required to spawn an entity
in the game screen.

A weapon entity known for both ranged and meele weapons as a
"projectile" is created using the hitbox components on the player layer
and a touch attack component on the enemy layer enabling the projetile
to contact eneimes. The projectile entity has a COmbat stats component
and this determines its damage.

Current melee weapons also have a Melee Component added to them within
the protile entity factory, this is responsible for controlling the
movmevenet of the projectile, (rotation and position) along with life
span.

Currently the project disposes of itself by reloacting itself to an
undisclosed location where it bides its time until the game is closed.
This is to be fixed