Reconfigured Weapon controller component to be more modular with different weapon designs to make use of a switch case based on weapon types to different function calls for each weapon to control movement.
Reconfigured Keyboard player input to use hash map to store if a key is
pressed and this is used for firing weapons on different keys + mouse 1.
Also holds a variable to to last known mouse location along with a
getter method for this variable.

Implemented Weapon Target component which is a component that holds a
weaopntype and an entity based on the weapon type and entity returns a
position to target.
-MOdified attack factory to make use of this new component.

Implemented new slingshot weapon that makes use of weapon target
component and the modified keyboard player input to have the projectile
home in on the mouse.